backend/requirements.txt
amqtt==0.11.1
python-dotenv==1.1.1